requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
aiomysql  
SQLAlchemy
passlib[bcrypt]
python-jose[cryptography]
openai